feat: add project data and default-selection option for side select box
