+++ subtasks/src/components/projects/ProjectTaskNavigation.vue
@@ -67,7 +67,36 @@
         />
       </svg>
     </button>
-
+    <div class="relative">
+      <button @click="showMenu = !showMenu">
+        <svg
+          class="w-6 h-6 text-gray-800 dark:text-white"
+          aria-hidden="true"
+          xmlns="http://www.w3.org/2000/svg"
+          width="24"
+          height="24"
+          fill="none"
+          viewBox="0 0 24 24"
+        >
+          <path
+            stroke="currentColor"
+            stroke-linecap="round"
+            stroke-width="2"
+            d="M12 6h.01M12 12h.01M12 18h.01"
+          />
+        </svg>
+      </button>
+      <div
+        :class="{ block: showMenu, hidden: !showMenu }"
+        class="absolute border border-gray-300 shadow-lg rounded-lg bg-white z-50 w-52 p-3"
+      >
+        <SideSelectBox
+          label-text="Select Project"
+          :default-value="projectId"
+          :options="projectOptions"
+        />
+      </div>
+    </div>
     <div class="flex flex-row items-center ml-auto">
       <div class="flex flex-row items-center">
         <RouterLink
@@ -141,10 +170,12 @@
 </template>
 
 <script setup lang="ts">
-import { computed, inject } from "vue";
+import { computed, inject, ref } from "vue";
 import { useQueryClient } from "@tanstack/vue-query";
+import SideSelectBox from "@components/shared/SideSelectBox.vue";
 import SearchBar from "@components/shared/SearchBar.vue";
 import { useTasks } from "../../composables/useTasks";
+import { useProjects } from "../../composables/useProjects";
 
 const {
   completed = false,
@@ -160,6 +191,7 @@ const {
   search?: string;
 }>();
 
+const showMenu = ref(false);
 const { data, isSuccess } = useTasks({
   projectId: () => projectId,
   completed: () => completed,
@@ -167,6 +199,18 @@ const { data, isSuccess } = useTasks({
   page: () => page,
   pageSize: () => pageSize,
 });
+
+const { data: projects } = useProjects({
+  completed: false,
+  page: 1,
+  pageSize: 100,
+});
+
+const projectOptions = computed(() =>
+  projects.value?.projects.map((project) => {
+    return { value: project.id, label: project.title };
+  }),
+);
 
 const projectsRemaining = computed(() => {
   if (!isSuccess) {
